pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "snapcast-monitor-kasa"
description = "Monitor a Snapcast server for client and stream state to control Kasa switches."
version = "1.0.0"
readme = "README.md"
authors = [
  {name = "Tucker Kern", email = "tuckkern@gmail.com"},
]
requires-python = ">=3.8"
license = {text = "MIT"}
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Topic :: Home Automation"
]
dependencies = [
  "snapcast",
  "python-kasa"
]

[project.urls]
Repository = "https://github.com/mill1000/snapcast-monitor"
Issues = "https://github.com/mill1000/snapcast-monitor/issues"

[project.scripts]
snapcast-monitor-kasa = "snapcast_monitor_kasa:main"

[tool.setuptools]
py-modules = ["snapcast_monitor_kasa"]
